[Fix] routers 파일 통합
